refactor: pacakge.json 타입스크립트 관련 devDep로 이동
