+++ package.json
@@ -8,8 +8,6 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.4",
     "axios": "^0.27.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
@@ -17,7 +15,6 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.0",
-    "typescript": "^4.6.4",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.4"
   },
@@ -71,8 +68,11 @@
     "@storybook/testing-library": "^0.0.11",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "@types/storybook__react": "^5.2.1",
     "@types/uuid": "^8.3.4",
+    "typescript": "^4.6.4",
     "babel-core": "^6.26.3",
     "babel-loader": "^8.2.5",
     "babel-plugin-emotion": "^11.0.0",
